exercise ready

+++ src/main/java/de/bhopp/ml/backend/ImageClassification.java
@@ -1,3 +1,3 @@
 package de.bhopp.ml.backend;
 
-public record ImageClassification(ImagenetLabel imagenetLabel, float probability) {}
+public record ImageClassification(ImagenetLabel imagenetLabel, double probability) {}
